pyproject.toml
[project]
name = "agentic-ai-identity"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=47.0.0",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "langchain-anthropic>=1.4.2",
    "langchain-core>=1.3.2",
    "langgraph>=1.1.10",
    "pyjwt[cryptography]>=2.12.1",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.46.0",
]
